+++ client/src/pages/Home/Home.js
@@ -1,0 +1,5 @@
+import React from "react";
+const Home = () => {
+    return (<p>This is the homepage</p>)
+}
+export default Home;
